+++ app/models/entity.rb
@@ -415,9 +415,9 @@ class Entity < ActiveRecord::Base
   }
   scope :by_relation_name, lambda {|relation_name|
     if relation_name
-      kind_ids = Relation.where(name: relation_name).map{|r| r.to_kind_ids}
-      kind_ids << Relation.where(reverse_name: relation_name).map{|r| r.from_kind_ids}
-      where(kind_id: kind_ids.flatten)
+      kind_ids = Relation.where(name: relation_name).map{|r| r.to_kind_id}
+      kind_ids << Relation.where(reverse_name: relation_name).map{|r| r.from_kind_id}
+      where(kind_id: kind_ids.flatten.uniq)
     else
       all
     end