+++ spec/controllers/relations_controller_spec.rb
@@ -14,14 +14,14 @@ RSpec.describe RelationsController, type: :controller do
     request.headers['accept'] = 'application/json'
   end
   
-  it "should save kind ids" do
-    @is_equivalent_to = Relation.find_by_name('is equivalent to')
+  it "should save kind id" do
+    is_equivalent_to = Relation.find_by_name('is equivalent to')
 
-    patch :update, id: @is_equivalent_to.id, relation: {
-      from_kind_ids: [@person_kind.id, @artwork_kind.id]
+    patch :update, id: is_equivalent_to.id, relation: {
+      from_kind_id: @artwork_kind.id
     }
     
-    expect(@is_equivalent_to.reload.from_kind_ids).to eql([ @person_kind.id, @artwork_kind.id ])
+    expect(is_equivalent_to.reload.from_kind_id).to eql(@artwork_kind.id)
   end
 
   it "should allow a json list without auth" do